requirements/base.txt
Django==3.2.11
django-crispy-forms==1.13.0
django-environ==0.8.1
django-extensions==3.1.5
mysqlclient==2.1.0
pip==21.3.1
Pillow==9.0.0
psycopg2-binary<=2.9
pytz
six==1.16.0
sorl-thumbnail==12.7.0
python-telegram-bot==13.14
requests==2.27.1
